make precompute_helmholtz a separate compilation unit note: hash() and hash_lut are needed in Helmholtz precomputation, we temporarily use forward declaration in precompute_helmholtz.cpp. For now, interaction_list.h is still only included in main.cpp.
