+++ src/helmholtz.cpp
@@ -7,28 +7,6 @@ namespace exafmm_t {
   real_t MU;
   std::vector<M2LData> M2Ldata;
 
-  // mixed-type gemm: A is complex_t matrix; B is real_t matrix
-  void gemm(int m, int n, int k, complex_t* A, real_t* B, complex_t* C) {
-    char transA = 'N', transB = 'N';
-    complex_t alpha(1., 0.), beta(0.,0.);
-#if FLOAT
-    scgemm_(&transA, &transB, &n, &m, &k, &alpha, B, &n, A, &k, &beta, C, &n);
-#else
-    dzgemm_(&transA, &transB, &n, &m, &k, &alpha, B, &n, A, &k, &beta, C, &n);
-#endif
-  }
-
-  // complex gemm by blas lib
-  void gemm(int m, int n, int k, complex_t* A, complex_t* B, complex_t* C) {
-    char transA = 'N', transB = 'N';
-    complex_t alpha(1., 0.), beta(0.,0.);
-#if FLOAT
-    cgemm_(&transA, &transB, &n, &m, &k, &alpha, B, &n, A, &k, &beta, C, &n);
-#else
-    zgemm_(&transA, &transB, &n, &m, &k, &alpha, B, &n, A, &k, &beta, C, &n);
-#endif
-  }
-
   // complex gemv by blas lib
   void gemv(int m, int n, complex_t* A, complex_t* x, complex_t* y) {
     char trans = 'T';
@@ -39,76 +17,6 @@ namespace exafmm_t {
 #else
     zgemv_(&trans, &n, &m, &alpha, A, &n, x, &incx, &beta, y, &incy);
 #endif
-  }
-
-  //! lapack svd with row major data: A = U*S*VT, A is m by n
-  void svd(int m, int n, real_t* A, real_t* S, real_t* U, real_t* VT) {
-    char JOBU = 'S', JOBVT = 'S';
-    int INFO;
-    int LWORK = std::max(3*std::min(m,n)+std::max(m,n), 5*std::min(m,n));
-    LWORK = std::max(LWORK, 1);
-    int k = std::min(m, n);
-    RealVec tS(k, 0.);
-    RealVec WORK(LWORK);
-#if FLOAT
-    sgesvd_(&JOBU, &JOBVT, &n, &m, A, &n, &tS[0], VT, &n, U, &k, &WORK[0], &LWORK, &INFO);
-#else
-    dgesvd_(&JOBU, &JOBVT, &n, &m, A, &n, &tS[0], VT, &n, U, &k, &WORK[0], &LWORK, &INFO);
-#endif
-    // copy singular values from 1d layout (tS) to 2d layout (S)
-    for(int i=0; i<k; i++) {
-      S[i*n+i] = tS[i];
-    }
-  }
-
-  void svd(int m, int n, complex_t* A, real_t* S, complex_t* U, complex_t* VT) {
-    char JOBU = 'S', JOBVT = 'S';
-    int INFO;
-    int LWORK = std::max(3*std::min(m,n)+std::max(m,n), 5*std::min(m,n));
-    LWORK = std::max(LWORK, 1);
-    int k = std::min(m, n);
-    RealVec tS(k, 0.);
-    ComplexVec WORK(LWORK);
-    RealVec RWORK(5*k);
-#if FLOAT
-    cgesvd_(&JOBU, &JOBVT, &n, &m, A, &n, &tS[0], VT, &n, U, &k, &WORK[0], &LWORK, &RWORK[0], &INFO);
-#else
-    zgesvd_(&JOBU, &JOBVT, &n, &m, A, &n, &tS[0], VT, &n, U, &k, &WORK[0], &LWORK, &RWORK[0], &INFO);
-#endif
-    // copy singular values from 1d layout (tS) to 2d layout (S)
-    for(int i=0; i<k; i++) {
-      S[i*n+i] = tS[i];
-    }
-  }
-
-  RealVec transpose(RealVec& vec, int m, int n) {
-    RealVec temp(vec.size());
-    for(int i=0; i<m; i++) {
-      for(int j=0; j<n; j++) {
-        temp[j*m+i] = vec[i*n+j];
-      }
-    }
-    return temp;
-  }
-
-  ComplexVec transpose(ComplexVec& vec, int m, int n) {
-    ComplexVec temp(vec.size());
-    for(int i=0; i<m; i++) {
-      for(int j=0; j<n; j++) {
-        temp[j*m+i] = vec[i*n+j];
-      }
-    }
-    return temp;
-  }
-
-  ComplexVec conjugate_transpose(ComplexVec& vec, int m, int n) {
-    ComplexVec temp(vec.size());
-    for(int i=0; i<m; i++) {
-      for(int j=0; j<n; j++) {
-        temp[j*m+i] = std::conj(vec[i*n+j]);
-      }
-    }
-    return temp;
   }
 
   void potential_P2P(RealVec& src_coord, ComplexVec& src_value, RealVec& trg_coord, ComplexVec& trg_value) {
